Move various functions into spots matching the backend request
