+++ src/services/activityService.js
@@ -37,9 +37,23 @@ const create = async (activityData) => {
     console.log(error)
   }
 }
+async function addToTrip(activity) {
+  try {
+    const res = await fetch(`${BASE_URL}/${activity._id}/activity-plan`, {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json',
+      'Authorization': `Bearer ${tokenService.getToken()}`},
+      body:JSON.stringify(activity),
+    })
+  return res.json()
+  } catch (err) {
+    throw err
+  }
+}
 
 export {
   index,
   show,
-  create
+  create,
+  addToTrip
 }
