+++ src/services/tripService.js
@@ -22,22 +22,10 @@ const show = async(id) => {
   }
 }
 
-async function addToTrip(activity) {
-  try {
-    const res = await fetch(`${BASE_URL}/addToTrip`, {
-      method: 'PUT',
-      headers: { 'Content-Type': 'application/json',
-      'Authorization': `Bearer ${tokenService.getToken()}`},
-      body:JSON.stringify(activity),
-    })
-  return res.json()
-  } catch (err) {
-    throw err
-  }
-}
+
 
 export {
   index,
   show,
-  addToTrip
+
 }
